--- src/actions/constants.js
@@ -1,19 +1,0 @@
-//constants
-export const ADD_BOOK = 'ADD_BOOK';
-export const CREATE_USER = 'CREATE_USER';
-export const EMPTY_CART = 'EMPTY_CART';
-export const FETCH_BOOKS = 'FETCH_BOOKS';
-export const GET_FAVORITES = 'GET_FAVORITES';
-export const GET_FEATURED = 'GET_FEATURED';
-export const GET_HISTORY = 'GET_HISTORY';
-export const GET_QUERY_LOG = 'GET_QUERY_LOG';
-export const GET_SUGGESTIONS = 'GET_SUGGESTIONS';
-export const NO_BOOKS = 'NO_BOOKS';
-export const RENDER_ERROR = 'RENDER_ERROR';
-export const REMOVE_ERROR = 'REMOVE_ERROR';
-export const SIGN_IN = 'SIGN_IN';
-export const SIGN_OUT = 'SIGN_OUT';
-export const UNWRAP = 'UNWRAP';
-export const SYNC_CART = 'SYNC_CART';
-export const LOADING = 'LOADING';
-export const FIREBASE = 'FIREBASE';
